+++ src/main/java/com/devbliss/doctest/LogicDocTest.java
@@ -23,7 +23,7 @@ import de.devbliss.apitester.ApiTest;
 import de.devbliss.apitester.ApiTest.HTTP_REQUEST;
 import de.devbliss.apitester.Cookie;
 
-public class LogicDocTest {
+public abstract class LogicDocTest {
 
     protected static DocTestMachine docTest;
 
@@ -33,7 +33,8 @@ public class LogicDocTest {
 
     @Before
     public void ensureDocTestClassSet() {
-        docTest.beginDoctest(this.getClass().getCanonicalName(), getIntroduction());
+        fileHelper.validateFileName(getFileName());
+        docTest.beginDoctest(getFileName(), getIntroduction());
     }
 
     @AfterClass
@@ -62,6 +63,13 @@ public class LogicDocTest {
     protected String getIntroduction() {
         return new String();
     }
+
+    /**
+     * Returns the name of the report file
+     * 
+     * @return name of the file without extension
+     */
+    protected abstract String getFileName();
 
     public void say(String say) {
         docTest.say(say);