+++ src/test/java/integration/RequestsIntegrationTest.java
@@ -43,6 +43,11 @@ public class RequestsIntegrationTest extends DocTest {
         DocTest.setApi(API);
     }
 
+    @Override
+    protected String getFileName() {
+        return "HttpRequests";
+    }
+
     private Object obj;
     private ApiResponse response;
     private URI uri;